fix: preserve native multipart snapshot authority and cleanup
